backend/requirements.txt
fastapi==0.116.0
uvicorn==0.25.0
pypdf2==3.0.1
faiss-cpu==1.12.0
numpy==2.3.1
python-multipart==0.0.20